ajuste de variavel int para float

+++ PythonExercicios/ex014.py
@@ -1,6 +1,6 @@
 # TODO Escreva um programa que converta uma temperatura digitando em graus Celsius e converta para graus Fahrenheit.
 
-c = int(input('Qual é a temperatura em Celsius?: '))
+c = float(input('Qual é a temperatura em Celsius?: '))
 
 f = (c * 9/5) + 32
